Переносит обработчики событий в компоненты
