+++ src/utils/common.js
@@ -1,0 +1,11 @@
+const getRandomInteger = (min, max) => {
+  return min + Math.floor(Math.random() * (max + 1 - min));
+};
+const getRandomItem = (arr) => {
+  return arr[Math.floor(Math.random() * arr.length)];
+};
+
+export {
+  getRandomInteger,
+  getRandomItem
+};
