+++ src/utils/render.js
@@ -26,18 +26,9 @@ const createElement = (template) => {
   return newElement.firstChild;
 };
 
-const getRandomInteger = (min, max) => {
-  return min + Math.floor(Math.random() * (max + 1 - min));
-};
-const getRandomItem = (arr) => {
-  return arr[Math.floor(Math.random() * arr.length)];
-};
-
 export {
   RenderPosition,
   render,
   renderTemplate,
   createElement,
-  getRandomInteger,
-  getRandomItem,
 };
